feat: better msi-init

+++ .config/fish/functions/msi-init.fish
@@ -1,21 +1,27 @@
 function msi-init
-    rich -a rounded -S blue -p "[blue]AWS SSO"
+    rich -a rounded -S blue --style blue -p "AWS SSO"
     aws-sso-util login --all
 
-    rich -a rounded -S blue -p "[blue]ScaleFt"
+    rich -a rounded -S blue --style blue -p "ScaleFt"
     sft login
 
-    rich -a rounded -S blue -p "[blue]MSI BitBucket"
-    rich --style italic -p "From [blue]https://confluence.mot-solutions.com/display/REMOT/Bitbucket+Access+Flow[/blue] (last updated [bold]19-Aug-2022[/bold])."
-    rich --style italic -p "Saved to [blue]s3://duybui-msi-shellscripts/msi-bitbucket.sh[/blue] in [blue]aws.aec.dev.duy[/blue] account."
-    aws --profile=sandbox s3 cp s3://duybui-msi-shellscripts/msi-bitbucket.sh - | bash -
+    rich -a rounded -S blue --style blue -p "MSI BitBucket"
+    if test -e ~/iap.header; and test (jq -R 'split(".") | .[1] | @base64d | fromjson | .exp' <~/iap.header) -gt (now)
+        set -l exp (math (jq -R 'split(".") | .[1] | @base64d | fromjson | .exp' <~/iap.header) - (now))
+        rich --style green -p "Session expires in $exp seconds."
+    else
+        rich --style blue -p "Invoking [bold italic]msi-bitbucket.sh[/bold italic]"
+        rich --style italic -p "From [blue]https://confluence.mot-solutions.com/display/REMOT/Bitbucket+Access+Flow[/blue] (last updated [bold]19-Aug-2022[/bold])."
+        rich --style italic -p "Saved to [blue]s3://duybui-msi-shellscripts/msi-bitbucket.sh[/blue] in [blue]aws.aec.dev.duy[/blue] account."
+        aws --profile=sandbox s3 cp s3://duybui-msi-shellscripts/msi-bitbucket.sh - | bash -
+    end
 
-    rich -a rounded -S blue -p "[blue]ActiveEye Gateway"
+    rich -a rounded -S blue --style blue -p "ActiveEye Gateway"
     for pod in hawk wasp lion orca wolf bear
         if ssh -o ConnectTimeout=5 -q gateway1.ec2.$pod.activeeye.com exit
-            rich -p "[green][bold italic]gateway1.ec2.$pod.activeeye.com[/bold italic] is up"
+            rich --style green -p "[bold italic]gateway1.ec2.$pod.activeeye.com[/bold italic] is accessible"
         else
-            rich -p "[blue]Invoking [bold italic]https://gateway1.$pod.activeeye.com/gateway"
+            rich --style blue -p "Invoking [bold italic]https://gateway1.$pod.activeeye.com/gateway[/bold italic]"
             xdg-open "https://gateway1.$pod.activeeye.com/gateway"
         end
     end
